Piece Structure now exists and has a lot of useful functions such as :
CreatePiece, RemovePiece, MovePiece

Also created a board generator that can convert board space to world space and generate pieces on the board.

Pieces hold references to their game objects. The visual needs to be updated when the piece moves in the piece structure, which is handlede by the MovePiece function,